feat: add basic terrain and cells (1/2)
Add a basic terrain with multiple painted textures.

Considered the first texture (grass) as being the only valid texture for constructing.

In the gridSystem, add a customGrid that will hold a new Cell class.

This Cell class will specify if a building can be constructed.

Use the alphamaps from the terrainData to know the repartition of textures on a specific cell and therefore compute if a building can be constructed on this cell